requirements.txt
jinja2>=3.1.2
pyyaml>=6.0.1
typing-extensions>=4.7.1
langchain>=0.1.0
langchain-anthropic>=0.1.1
python-dotenv>=1.0.0
pydantic>=2.5.2
anthropic>=0.18.1
streamlit==1.26.0
streamlit-ace>=0.1.1
watchdog>=3.0.0 